requirements.txt
flask
flask-cors
spacy
pdfplumber
pytesseract
python-docx
werkzeug
